Refactor: Move classes
